Remove copy to clipboard button
Clipboard is disabled in insecure contexts (HTTP) :(

This is yet another thorn in our side stemming from running on the LAN
(the other thorn is mDNS being unreliable). I think eventually we will
want to host the frontend on the public internet and put it behind RC
OAuth. This would solve both of these problems, and also make it
possible to print to the hub remotely (unclear if desirable).